Preserve UFC shard model compatibility

+++ products/feature-flagging/feature-flagging-bootstrap/src/main/java/datadog/trace/api/featureflag/ufc/v1/Shard.java
@@ -2,12 +2,19 @@ package datadog.trace.api.featureflag.ufc.v1;
 
 import java.util.List;
 
+/**
+ * A UFC shard.
+ *
+ * <p>{@code totalShards} stores the unsigned 32-bit wire value in an {@code int} to preserve this
+ * bootstrap model's binary compatibility. Consumers must use {@link Integer#toUnsignedLong(int)}
+ * before comparing it or using it for arithmetic.
+ */
 public class Shard {
   public final String salt;
   public final List<ShardRange> ranges;
-  public final long totalShards;
+  public final int totalShards;
 
-  public Shard(final String salt, final List<ShardRange> ranges, final long totalShards) {
+  public Shard(final String salt, final List<ShardRange> ranges, final int totalShards) {
     this.salt = salt;
     this.ranges = ranges;
     this.totalShards = totalShards;
