+++ products/feature-flagging/feature-flagging-bootstrap/src/main/java/datadog/trace/api/featureflag/ufc/v1/ShardRange.java
@@ -1,10 +1,17 @@
 package datadog.trace.api.featureflag.ufc.v1;
 
+/**
+ * A UFC shard range.
+ *
+ * <p>{@code start} and {@code end} store unsigned 32-bit wire values in {@code int} fields to
+ * preserve this bootstrap model's binary compatibility. Consumers must use {@link
+ * Integer#toUnsignedLong(int)} before comparing them.
+ */
 public class ShardRange {
-  public final long start;
-  public final long end;
+  public final int start;
+  public final int end;
 
-  public ShardRange(final long start, final long end) {
+  public ShardRange(final int start, final int end) {
     this.start = start;
     this.end = end;
   }
